+++ src/Makefile
@@ -1,2 +1,2 @@
 all:
-	g++ main.cpp pos.cpp utilities.cpp castling.cpp move.cpp move_gen.cpp test.cpp timer.cpp search.cpp --std=c++0x -O3
+	g++ main.cpp pos.cpp utilities.cpp castling.cpp move.cpp move_gen.cpp test.cpp timer.cpp search.cpp evaluation.cpp --std=c++0x -O3
