+++ src/evaluation.cpp
@@ -1,0 +1,22 @@
+#include "evaluation.h"
+
+int count_material(Pos &pos, Color side)
+{
+    int material = 0;
+    for (int i = 0; i < 6; i++)
+    {
+        material += __builtin_popcountll(pos.pieces_bbs[i] & pos.colors_bbs[side]) * get_piece_value(Piece(i));
+    }
+    return material;
+}
+
+int eval(Pos &pos)
+{
+    Color side = pos.turn;
+    Color enemy = Color(!side);
+
+    int current_player_material = count_material(pos, side);
+    int enemy_player_material = count_material(pos, enemy);
+
+    return current_player_material - enemy_player_material;
+}
