+++ src/evaluation.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "types.h"
+#include "pos.h"
+
+int count_material(Pos &pos, Color side);
+
+int eval(Pos& pos);
