+++ src/test.h
@@ -1,6 +1,3 @@
 #pragma once
-#include "pos.h"
-#include "move_gen.h"
-#include "timer.h"
 
 void test_fens();
